interrupted sync changes
